backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.22
psycopg[binary]==3.2.8
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
requests==2.32.3
pydantic==2.8.0
Pillow==10.4.0
supabase==2.7.4 